Move some logic from LocalController to new model
